requirements.txt
certifi==2017.11.5
chardet==3.0.4
google-api-python-client==1.6.4
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
olefile==0.44
Pillow==4.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.5.3
PySocks==1.6.8
pytesseract==0.1.7
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
